requirements.txt
streamlit
pandas
google-cloud-bigquery
google-cloud-storage
google-api-core
db-dtypes
pyarrow
google-cloud-aiplatform
google-genai
python-dotenv
